sync-server: move usage monitoring from project to new package
