--- src/packages/project/project-status/index.ts
@@ -1,6 +1,0 @@
-/*
- *  This file is part of CoCalc: Copyright © 2020 Sagemath, Inc.
- *  License: MS-RSL – see LICENSE.md for details
- */
-
-export { get_ProjectStatusServer, ProjectStatusServer } from "./server";
